refactor: rename type to _types to avoid same name libraries
