+++ cellacdc/apps.py
@@ -79,7 +79,7 @@ from . import widgets
 from . import user_profile_path
 from . import features
 from . import _core
-from . import types
+from . import _types
 from . import plot
 from . import urls
 from .acdc_regex import float_regex
@@ -10335,7 +10335,7 @@ class FunctionParamsDialog(QBaseDialog):
         ArgsWidgets_list = []
         
         for row, ArgSpec in enumerate(params_argspecs):
-            if types.is_widget_not_required(ArgSpec):
+            if _types.is_widget_not_required(ArgSpec):
                 continue
             
             arg_name = ArgSpec.name         
@@ -10352,14 +10352,14 @@ class FunctionParamsDialog(QBaseDialog):
             
             isVectorEntry = False
             try:
-                if isinstance(ArgSpec.type(), types.Vector):
+                if isinstance(ArgSpec.type(), _types.Vector):
                     isVectorEntry = True
             except Exception as err:
                 pass
             
             isFolderPath = False
             try:
-                if isinstance(ArgSpec.type(), types.FolderPath):
+                if isinstance(ArgSpec.type(), _types.FolderPath):
                     isFolderPath = True
             except Exception as err:
                 pass
@@ -10447,7 +10447,7 @@ class FunctionParamsDialog(QBaseDialog):
                 widget.sigValueChanged.connect(self.emitValuesChanged)
             elif isCustomListType:
                 items = ArgSpec.type().values
-                ArgSpec.type.cast_dtype = types.to_str
+                ArgSpec.type.cast_dtype = _types.to_str
                 defaultVal = str(ArgSpec.default)
                 combobox = widgets.AlphaNumericComboBox()
                 combobox.addItems(items)
@@ -10870,7 +10870,7 @@ class QDialogModelParams(QDialog):
         isDualChannelModel = (
             self.model_name.find('cellpose') != -1 
             or any([
-                types.is_second_channel_type(ArgSpec.type) 
+                _types.is_second_channel_type(ArgSpec.type) 
                 for ArgSpec in ArgSpecs_list
             ])
         )
@@ -10891,10 +10891,10 @@ class QDialogModelParams(QDialog):
             start_row += 1
         
         for row, ArgSpec in enumerate(ArgSpecs_list):
-            if types.is_second_channel_type(ArgSpec.type):
+            if _types.is_second_channel_type(ArgSpec.type):
                 continue
             
-            if types.is_widget_not_required(ArgSpec):
+            if _types.is_widget_not_required(ArgSpec):
                 continue
             
             row = row + start_row
@@ -10918,14 +10918,14 @@ class QDialogModelParams(QDialog):
             
             isVectorEntry = False
             try:
-                if isinstance(ArgSpec.type(), types.Vector):
+                if isinstance(ArgSpec.type(), _types.Vector):
                     isVectorEntry = True
             except Exception as err:
                 pass
             
             isFolderPath = False
             try:
-                if isinstance(ArgSpec.type(), types.FolderPath):
+                if isinstance(ArgSpec.type(), _types.FolderPath):
                     isFolderPath = True
             except Exception as err:
                 pass
